Snapshot 91.0.4472.19
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2842152
Bot-Commit: Chrome Release Bot (LUCI) <chrome-official-brancher@chops-service-accounts.iam.gserviceaccount.com>
Cr-Commit-Position: refs/branch-heads/4472@{#288}
Cr-Branched-From: 3d60439cfb36485e76a1c5bb7f513d3721b20da1-refs/heads/master@{#870763}

+++ src/base/allocator/partition_allocator/memory_reclaimer.cc
@@ -6,7 +6,8 @@
 
 #include "base/allocator/partition_allocator/partition_alloc.h"
 #include "base/allocator/partition_allocator/partition_alloc_check.h"
-#include "base/allocator/partition_allocator/pcscan.h"
+#include "base/allocator/partition_allocator/partition_alloc_config.h"
+#include "base/allocator/partition_allocator/starscan/pcscan.h"
 #include "base/bind.h"
 #include "base/location.h"
 #include "base/metrics/histogram_functions.h"
@@ -106,7 +107,7 @@ PartitionAllocMemoryReclaimer::~PartitionAllocMemoryReclaimer() = default;
 void PartitionAllocMemoryReclaimer::ReclaimAll() {
   constexpr int kFlags = PartitionPurgeDecommitEmptySlotSpans |
                          PartitionPurgeDiscardUnusedSystemPages |
-                         PartitionPurgeForceAllFreed;
+                         PartitionPurgeAggressiveReclaim;
   Reclaim(kFlags);
 }
 
@@ -130,15 +131,19 @@ void PartitionAllocMemoryReclaimer::Reclaim(int flags) {
   // Lastly decommit empty slot spans and lastly try to discard unused pages at
   // the end of the remaining active slots.
   {
-    using PCScan = internal::PCScan<internal::ThreadSafe>;
-    const auto invocation_mode = flags & PartitionPurgeForceAllFreed
+    using PCScan = internal::PCScan;
+    const auto invocation_mode = flags & PartitionPurgeAggressiveReclaim
                                      ? PCScan::InvocationMode::kForcedBlocking
                                      : PCScan::InvocationMode::kBlocking;
     PCScan::Instance().PerformScanIfNeeded(invocation_mode);
   }
 
 #if defined(PA_THREAD_CACHE_SUPPORTED)
-  internal::ThreadCacheRegistry::Instance().PurgeAll();
+  // Don't completely empty the thread cache outside of low memory situations,
+  // as there is periodic purge which makes sure that it doesn't take too much
+  // space.
+  if (PartitionPurgeAggressiveReclaim)
+    internal::ThreadCacheRegistry::Instance().PurgeAll();
 #endif
 
   for (auto* partition : thread_safe_partitions_)
