+++ src/chrome/browser/ui/webui/signin/inline_login_handler_impl.cc
@@ -19,7 +19,6 @@
 #include "base/metrics/user_metrics.h"
 #include "base/metrics/user_metrics_action.h"
 #include "base/single_thread_task_runner.h"
-#include "base/strings/string16.h"
 #include "base/strings/string_number_conversions.h"
 #include "base/strings/string_util.h"
 #include "base/strings/utf_string_conversions.h"
@@ -37,9 +36,10 @@
 #include "chrome/browser/signin/account_consistency_mode_manager.h"
 #include "chrome/browser/signin/chrome_device_id_helper.h"
 #include "chrome/browser/signin/identity_manager_factory.h"
-#include "chrome/browser/signin/local_auth.h"
 #include "chrome/browser/signin/signin_promo.h"
 #include "chrome/browser/signin/signin_util.h"
+#include "chrome/browser/themes/theme_service.h"
+#include "chrome/browser/themes/theme_service_factory.h"
 #include "chrome/browser/ui/browser_commands.h"
 #include "chrome/browser/ui/browser_finder.h"
 #include "chrome/browser/ui/browser_list.h"
@@ -93,7 +93,6 @@ namespace {
 // InlineLoginHandlerImpl.
 enum class HandlerSigninReason {
   FORCED_SIGNIN_PRIMARY_ACCOUNT,
-  UNLOCK,
   FETCH_LST_ONLY
 };
 
@@ -102,28 +101,13 @@ HandlerSigninReason GetHandlerSigninReason(const GURL& url) {
   signin_metrics::Reason reason =
       signin::GetSigninReasonForEmbeddedPromoURL(url);
   switch (reason) {
-    case signin_metrics::Reason::REASON_FORCED_SIGNIN_PRIMARY_ACCOUNT:
+    case signin_metrics::Reason::kForcedSigninPrimaryAccount:
       return HandlerSigninReason::FORCED_SIGNIN_PRIMARY_ACCOUNT;
-    case signin_metrics::Reason::REASON_UNLOCK:
-      return HandlerSigninReason::UNLOCK;
-    case signin_metrics::Reason::REASON_FETCH_LST_ONLY:
+    case signin_metrics::Reason::kFetchLstOnly:
       return HandlerSigninReason::FETCH_LST_ONLY;
     default:
       NOTREACHED() << "Unexpected signin reason: " << static_cast<int>(reason);
       return HandlerSigninReason::FORCED_SIGNIN_PRIMARY_ACCOUNT;
-  }
-}
-
-// Converts from HandlerSigninReason to signin_metrics::Reason.
-signin_metrics::Reason GetSigninReasonFromHandlerSigninReason(
-    HandlerSigninReason r) {
-  switch (r) {
-    case HandlerSigninReason::FORCED_SIGNIN_PRIMARY_ACCOUNT:
-      return signin_metrics::Reason::REASON_FORCED_SIGNIN_PRIMARY_ACCOUNT;
-    case HandlerSigninReason::UNLOCK:
-      return signin_metrics::Reason::REASON_UNLOCK;
-    case HandlerSigninReason::FETCH_LST_ONLY:
-      return signin_metrics::Reason::REASON_FETCH_LST_ONLY;
   }
 }
 
@@ -237,7 +221,7 @@ void SetProfileLocked(const base::FilePath profile_path, bool locked) {
 }
 
 void SetProfileName(const base::FilePath& profile_path,
-                    const base::string16 name) {
+                    const std::u16string name) {
   ProfileAttributesEntry* entry =
       g_browser_process->profile_manager()
           ->GetProfileAttributesStorage()
@@ -281,9 +265,6 @@ void OnSyncSetupComplete(Profile* profile,
         /*is_primary_account_=*/true,
         password_manager::metrics_util::GaiaPasswordHashChange::
             SAVED_ON_CHROME_SIGNIN);
-
-    if (profiles::IsLockAvailable(profile))
-      LocalAuth::SetLocalAuthCredentials(profile, password);
   }
 
   if (has_primary_account && is_force_sign_in_with_usermanager &&
@@ -293,7 +274,7 @@ void OnSyncSetupComplete(Profile* profile,
     base::Optional<AccountInfo> primary_account_info =
         identity_manager->FindExtendedAccountInfoForAccountWithRefreshToken(
             primary_account);
-    base::string16 profile_name;
+    std::u16string profile_name;
     if (primary_account_info.has_value()) {
       profile_name =
           profiles::GetDefaultNameForNewSignedInProfile(*primary_account_info);
@@ -304,7 +285,10 @@ void OnSyncSetupComplete(Profile* profile,
     }
     SetProfileName(profile->GetPath(), profile_name);
     Browser* browser = chrome::FindBrowserWithProfile(profile);
-    if (browser) {
+
+    // Don't show the customization bubble if a valid policy theme is set.
+    if (browser &&
+        !ThemeServiceFactory::GetForProfile(profile)->UsingPolicyTheme()) {
       ApplyProfileColorAndShowCustomizationBubbleWhenNoValueSynced(
           browser, GenerateNewProfileColor().color);
     }
@@ -408,11 +392,6 @@ void InlineSigninHelper::OnClientOAuthSuccessAndBrowserOpened(
   std::string primary_email =
       identity_manager->GetPrimaryAccountInfo(signin::ConsentLevel::kSync)
           .email;
-  if (gaia::AreEmailsSame(email_, primary_email) &&
-      reason == HandlerSigninReason::UNLOCK && !password_.empty() &&
-      profiles::IsLockAvailable(profile_)) {
-    LocalAuth::SetLocalAuthCredentials(profile_, password_);
-  }
 
   if (!password_.empty()) {
     scoped_refptr<password_manager::PasswordStore> password_store =
@@ -426,44 +405,19 @@ void InlineSigninHelper::OnClientOAuthSuccessAndBrowserOpened(
     }
   }
 
-  if (reason == HandlerSigninReason::UNLOCK) {
-    DCHECK(!identity_manager->GetPrimaryAccountInfo(signin::ConsentLevel::kSync)
-                .IsEmpty());
-
-    identity_manager->GetAccountsMutator()->AddOrUpdateAccount(
-        gaia_id_, email_, result.refresh_token,
-        result.is_under_advanced_protection,
-        signin_metrics::SourceForRefreshTokenOperation::
-            kInlineLoginHandler_Signin);
-
-    if (signin::IsAutoCloseEnabledInEmbeddedURL(current_url_)) {
-      // Close the gaia sign in tab via a task to make sure we aren't in the
-      // middle of any webui handler code.
-      base::ThreadTaskRunnerHandle::Get()->PostTask(
-          FROM_HERE,
-          base::BindOnce(&InlineLoginHandlerImpl::CloseTab, handler_));
-    }
-
-    identity_manager->GetAccountsCookieMutator()->AddAccountToCookie(
-        identity_manager->GetPrimaryAccountId(signin::ConsentLevel::kSync),
-        gaia::GaiaSource::kPrimaryAccountManager, {});
-
-    signin_metrics::LogSigninReason(
-        GetSigninReasonFromHandlerSigninReason(reason));
-  } else {
-    if (confirm_untrusted_signin_) {
-      // Display a confirmation dialog to the user.
-      base::RecordAction(
-          base::UserMetricsAction("Signin_Show_UntrustedSigninPrompt"));
-      Browser* browser = chrome::FindLastActiveWithProfile(profile_);
-      browser->window()->ShowOneClickSigninConfirmation(
-          base::UTF8ToUTF16(email_),
-          base::BindOnce(&InlineSigninHelper::UntrustedSigninConfirmed,
-                         base::Unretained(this), result.refresh_token));
-      return;
-    }
-    CreateSyncStarter(result.refresh_token);
+  if (confirm_untrusted_signin_) {
+    // Display a confirmation dialog to the user.
+    base::RecordAction(
+        base::UserMetricsAction("Signin_Show_UntrustedSigninPrompt"));
+    Browser* browser = chrome::FindLastActiveWithProfile(profile_);
+    browser->window()->ShowOneClickSigninConfirmation(
+        base::UTF8ToUTF16(email_),
+        base::BindOnce(&InlineSigninHelper::UntrustedSigninConfirmed,
+                       base::Unretained(this), result.refresh_token));
+    return;
   }
+  CreateSyncStarter(result.refresh_token);
+
   base::ThreadTaskRunnerHandle::Get()->DeleteSoon(FROM_HERE, this);
 }
 
@@ -517,8 +471,10 @@ void InlineSigninHelper::CreateSyncStarter(const std::string& refresh_token) {
 
 void InlineSigninHelper::OnClientOAuthFailure(
     const GoogleServiceAuthError& error) {
-  if (handler_)
-    handler_->HandleLoginError(error.ToString(), base::string16());
+  if (handler_) {
+    handler_->HandleLoginError(
+        SigninUIError::FromGoogleServiceAuthError(email_, error));
+  }
 
   HandlerSigninReason reason = GetHandlerSigninReason(current_url_);
   if (reason != HandlerSigninReason::FETCH_LST_ONLY) {
@@ -597,9 +553,6 @@ void InlineLoginHandlerImpl::SetExtraInitParams(base::DictionaryValue& params) {
 
   std::string flow;
   switch (reason) {
-    case HandlerSigninReason::UNLOCK:
-      flow = "reauth";
-      break;
     case HandlerSigninReason::FORCED_SIGNIN_PRIMARY_ACCOUNT:
       flow = "enterprisefsi";
       break;
@@ -672,8 +625,7 @@ void InlineLoginHandlerImpl::CompleteLogin(const std::string& email,
     return;
   }
 
-  DCHECK(reason == HandlerSigninReason::FORCED_SIGNIN_PRIMARY_ACCOUNT ||
-         reason == HandlerSigninReason::UNLOCK);
+  DCHECK(reason == HandlerSigninReason::FORCED_SIGNIN_PRIMARY_ACCOUNT);
   DCHECK(signin_util::IsForceSigninEnabled());
 
   ProfileManager* manager = g_browser_process->profile_manager();
@@ -750,8 +702,9 @@ void InlineLoginHandlerImpl::FinishCompleteLogin(
     credential_provider::UiExitCodes exit_code = ValidateSigninEmail(
         validate_gaia_id, email_domains, params.email, params.gaia_id);
     if (exit_code != credential_provider::kUiecSuccess) {
-      params.handler->HandleLoginError(base::NumberToString((int)exit_code),
-                                       base::UTF8ToUTF16(params.email));
+      params.handler->HandleLoginError(
+          SigninUIError::FromCredentialProviderUiExitCode(params.email,
+                                                          exit_code));
       return;
     } else {
       // Validation has already been done for GCPW, so clear the validate
@@ -769,9 +722,7 @@ void InlineLoginHandlerImpl::FinishCompleteLogin(
   if (validate_email == "1" && !default_email.empty()) {
     if (!gaia::AreEmailsSame(params.email, default_email)) {
       params.handler->HandleLoginError(
-          l10n_util::GetStringFUTF8(IDS_SYNC_WRONG_EMAIL,
-                                    base::UTF8ToUTF16(default_email)),
-          base::UTF8ToUTF16(params.email));
+          SigninUIError::WrongReauthAccount(params.email, default_email));
       return;
     }
   }
@@ -786,25 +737,12 @@ void InlineLoginHandlerImpl::FinishCompleteLogin(
                         ? signin_metrics::HISTOGRAM_WITH_ADVANCED
                         : signin_metrics::HISTOGRAM_WITH_DEFAULTS);
 
-  CanOfferSigninType can_offer_for = CAN_OFFER_SIGNIN_FOR_ALL_ACCOUNTS;
-  if (reason == HandlerSigninReason::UNLOCK) {
-    std::string primary_username =
-        IdentityManagerFactory::GetForProfile(profile)
-            ->GetPrimaryAccountInfo(signin::ConsentLevel::kSync)
-            .email;
-    if (!gaia::AreEmailsSame(default_email, primary_username))
-      can_offer_for = CAN_OFFER_SIGNIN_FOR_SECONDARY_ACCOUNT;
-  }
-
   SigninUIError can_offer_error = SigninUIError::Ok();
-  if (reason != HandlerSigninReason::FETCH_LST_ONLY) {
-    can_offer_error =
-        CanOfferSignin(profile, can_offer_for, params.gaia_id, params.email);
-  }
+  if (reason != HandlerSigninReason::FETCH_LST_ONLY)
+    can_offer_error = CanOfferSignin(profile, params.gaia_id, params.email);
+
   if (!can_offer_error.IsOk()) {
-    params.handler->HandleLoginError(
-        base::UTF16ToUTF8(can_offer_error.message()),
-        base::UTF8ToUTF16(params.email));
+    params.handler->HandleLoginError(can_offer_error);
     return;
   }
 
@@ -833,8 +771,7 @@ void InlineLoginHandlerImpl::FinishCompleteLogin(
   }
 }
 
-void InlineLoginHandlerImpl::HandleLoginError(const std::string& error_msg,
-                                              const base::string16& email) {
+void InlineLoginHandlerImpl::HandleLoginError(const SigninUIError& error) {
   content::WebContents* contents = web_ui()->GetWebContents();
   const GURL& current_url = contents->GetURL();
   HandlerSigninReason reason = GetHandlerSigninReason(current_url);
@@ -842,11 +779,12 @@ void InlineLoginHandlerImpl::HandleLoginError(const std::string& error_msg,
   if (reason == HandlerSigninReason::FETCH_LST_ONLY) {
     base::Value error_value(base::Value::Type::DICTIONARY);
 #if defined(OS_WIN)
-    // If the message is an integer error code, send it as part of the result.
-    int exit_code = 0;
-    if (base::StringToInt(error_msg, &exit_code)) {
+    // If the error contains an integer error code, send it as part of the
+    // result.
+    if (error.type() ==
+        SigninUIError::Type::kFromCredentialProviderUiExitCode) {
       error_value.SetKey(credential_provider::kKeyExitCode,
-                         base::Value(exit_code));
+                         base::Value(error.credential_provider_exit_code()));
     }
 #endif
     SendLSTFetchResultsMessage(error_value);
@@ -859,9 +797,9 @@ void InlineLoginHandlerImpl::HandleLoginError(const std::string& error_msg,
   if (profile->IsSystemProfile())
     profile = g_browser_process->profile_manager()->GetProfileByPath(
         ProfilePicker::GetForceSigninProfilePath());
-  if (!error_msg.empty()) {
-    LoginUIServiceFactory::GetForProfile(profile)->DisplayLoginResult(
-        browser, base::UTF8ToUTF16(error_msg), email);
+  if (!error.IsOk()) {
+    LoginUIServiceFactory::GetForProfile(profile)->DisplayLoginResult(browser,
+                                                                      error);
   }
 }
